+++ src/main/java/com/ayakacraft/carpetayakaaddition/mixin/utils/LivingEntityAccessor.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.ayakacraft.carpetayakaaddition.mixin.commands.gohome;
+package com.ayakacraft.carpetayakaaddition.mixin.utils;
 
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;