+++ changelog.md
@@ -1,5 +1,7 @@
 ### General
 
 - Added carpet rule `accurateDispenser`
-- Support for Minecraft 26.1-snapshot-1
+- Added carpet rule `commandEndermanBlockList` and command `/endermanBlockList`
+  - See https://github.com/AyakaCraft/Carpet-Ayaka-Addition/blob/main/README.md#commands for more information
+- Support for Minecraft 26.1-snapshot-2
 - Requires Fabric Loader 0.18.3+